Silly fixes (#1251)
* Start of silly fixes

* Fixes #1249

* Fixes #1155

* Add a comments section to the website

* Fix #821, fix #1094

* Fix #1093, fix #960

* Fix #1070

* Fix #1091

* Review comment, apply NativeAOT fix only on NS21+

* Add support for Vulkan specification specialization (SC), full regen

* Fix build

* Last silly fix

* Calm down

* Address review comments

* Build fixes

* How did that not work...

* Fix regen CI, only add None enum if there isn't a 0, update patch notes

* Include the pushable token for CI

* Don't remove the Flags attribute where unnecessary

* Remove problematic type, remove erroneous additions of None in enums

* New binaries for Assimp on Darwin 21.6.0 Darwin Kernel Version 21.6.0: Mon Dec 19 20:44:01 PST 2022; root:xnu-8020.240.18~2/RELEASE_X86_64 (#1310)


* Regenerate bindings as of 10/03/2023 (#1324)

* Regenerated bindings as of 10/03/2023 (16:54:01)

* Regenerated bindings as of 10/03/2023 (17:35:57)

* Regenerated bindings as of 10/03/2023 (18:27:27)

---------


* Build fixes

* Review comments, update WebGPU?

* Fix Vulkan build errors

---------

+++ src/Vulkan/Silk.NET.Vulkan/Structs/VideoEncodeH265RateControlLayerInfoEXT.gen.cs
@@ -23,7 +23,7 @@ namespace Silk.NET.Vulkan
         (
             StructureType? sType = StructureType.VideoEncodeH265RateControlLayerInfoExt,
             void* pNext = null,
-            byte? temporalId = null,
+            uint? temporalId = null,
             Bool32? useInitialRcQp = null,
             VideoEncodeH265QpEXT? initialRcQp = null,
             Bool32? useMinQp = null,
@@ -101,10 +101,10 @@ namespace Silk.NET.Vulkan
         [NativeName("Name", "pNext")]
         public void* PNext;
 /// <summary></summary>
-        [NativeName("Type", "uint8_t")]
-        [NativeName("Type.Name", "uint8_t")]
+        [NativeName("Type", "uint32_t")]
+        [NativeName("Type.Name", "uint32_t")]
         [NativeName("Name", "temporalId")]
-        public byte TemporalId;
+        public uint TemporalId;
 /// <summary></summary>
         [NativeName("Type", "VkBool32")]
         [NativeName("Type.Name", "VkBool32")]
